+++ src/data_item.rs
@@ -742,3 +742,20 @@ pub struct Mode2CodeOctalRepresentation {
 impl Mode2CodeOctalRepresentation {
     pub const FRN_48: u8 = 0b10_0000;
 }
+
+/// Confidence level for each bit of a Mode-1 reply as provided by
+/// a monopulse SSR station
+///
+/// Data Item I048/065
+#[derive(Debug, PartialEq, DekuRead, DekuWrite)]
+#[deku(ctx = "_: deku::ctx::Endian")]
+pub struct Mode1CodeConfidenceIndicator {
+    #[deku(bits = "3", endian = "big")]
+    pub spare: u8,
+    #[deku(bits = "5", endian = "big")]
+    pub data: u8,
+}
+
+impl Mode1CodeConfidenceIndicator {
+    pub const FRN_48: u8 = 0b1_0000;
+}