+++ src/fourty_eight.rs
@@ -4,8 +4,8 @@ use crate::data_item::{
     ACASResolutionAdvisoryReport, AircraftAddress, AircraftIdentification,
     CalculatedPositionCartesianCorr, CalculatedTrackVelocity, CommunicationsCapabilityFlightStatus,
     DataSourceIdentifier, FlightLevelInBinaryRepresentation, HeightMeasuredBy3dRadar,
-    MeasuredPositionInPolarCoordinates, Mode1CodeOctalRepresentation, Mode2CodeOctalRepresentation,
-    Mode3ACodeConfidenceIndicator, Mode3ACodeInOctalRepresentation,
+    MeasuredPositionInPolarCoordinates, Mode1CodeConfidenceIndicator, Mode1CodeOctalRepresentation,
+    Mode2CodeOctalRepresentation, Mode3ACodeConfidenceIndicator, Mode3ACodeInOctalRepresentation,
     ModeCCodeAndConfidenceIndicator, ModeSMBData, RadarPlotCharacteristics, RadialDopplerSpeed,
     TargetReportDescriptor, TimeOfDay, TrackNumber, TrackQuality, TrackStatus,
     WarningErrorConditionsTargetClass,
@@ -122,7 +122,12 @@ pub struct Cat48 {
         cond = "is_fspec(Mode2CodeOctalRepresentation::FRN_48, fspec, 3)"
     )]
     pub mode_2_code_octal_representation: Option<Mode2CodeOctalRepresentation>,
-    // FRN 23
+    /// FRN 23
+    #[deku(
+        skip,
+        cond = "is_fspec(Mode1CodeConfidenceIndicator::FRN_48, fspec, 3)"
+    )]
+    pub mode_1_code_confidence: Option<Mode1CodeConfidenceIndicator>,
     // FRN 24
     // FRN 25
     // FRN 26