+++ tests/public_api.rs
@@ -658,3 +658,17 @@ fn test_mode2code_octal_representation() {
         assert_eq!(field.data, 0x01);
     }
 }
+
+#[test]
+fn test_mode1codeconfidence() {
+    let bytes = vec![0x30, 0x00, 0x08, 0x01, 0x01, 0x01, 0b0001_0000, 0x01];
+    let (_, packet) = AsterixPacket::from_bytes((&bytes, 0)).unwrap();
+    assert_eq!(packet.to_bytes().unwrap(), bytes);
+
+    if let AsterixMessage::Cat48(ref message) = packet.messages[0] {
+        assert_eq!(message.fspec, &[0x01, 0x01, 0x01, 0b0001_0000]);
+
+        let field = message.mode_1_code_confidence.as_ref().unwrap();
+        assert_eq!(field.data, 0x01);
+    }
+}